+++ dev/setup_secrets.sh
@@ -8,7 +8,7 @@ TODO: These instructions are currently pieced together from old disparate
 instances, and are not yet fully organized.
 
 The original template file should be:
-~/misc/templates/PYPKG/dev/setup_secrets.sh
+~/code/xcookie/dev/setup_secrets.sh
 
 Development script for updating secrets when they rotate
 
@@ -39,7 +39,7 @@ GITLAB ACTION INSTRUCTIONS
     ```bash
     cat .setup_secrets.sh | \
         sed "s|utils|<YOUR-GROUP>|g" | \
-        sed "s|PYPKG|<YOUR-REPO>|g" | \
+        sed "s|xcookie|<YOUR-REPO>|g" | \
         sed "s|travis-ci-Erotemic|<YOUR-GPG-ID>|g" | \
         sed "s|CI_SECRET|<YOUR_CI_SECRET>|g" | \
         sed "s|GITLAB_ORG_PUSH_TOKEN|<YOUR_GIT_ORG_PUSH_TOKEN>|g" | \
@@ -48,12 +48,12 @@ GITLAB ACTION INSTRUCTIONS
     ```
 
     * Make sure you add Runners to your project 
-    https://gitlab.org.com/utils/PYPKG/-/settings/ci_cd 
+    https://gitlab.org.com/utils/xcookie/-/settings/ci_cd 
     in Runners-> Shared Runners
     and Runners-> Available specific runners
 
     * Ensure that you are auto-cancel redundant pipelines.
-    Navigate to https://gitlab.kitware.com/utils/PYPKGS/-/settings/ci_cd and ensure "Auto-cancel redundant pipelines" is checked.
+    Navigate to https://gitlab.kitware.com/utils/xcookie/-/settings/ci_cd and ensure "Auto-cancel redundant pipelines" is checked.
 
     More details are here https://docs.gitlab.com/ee/ci/pipelines/settings.html#auto-cancel-redundant-pipelines
 
@@ -118,14 +118,16 @@ setup_package_environs(){
     echo "Choose an organization specific setting or make your own. This needs to be generalized more"
 }
 
+### FIXME: Should be configurable for general use
+
 setup_package_environs_gitlab_kitware(){
     echo '
     export VARNAME_CI_SECRET="CI_KITWARE_SECRET"
-    export VARNAME_TWINE_USERNAME="TWINE_USERNAME"
-    export VARNAME_TWINE_PASSWORD="TWINE_PASSWORD"
-    export VARNAME_TEST_TWINE_USERNAME="TEST_TWINE_USERNAME"
-    export VARNAME_TEST_TWINE_PASSWORD="TEST_TWINE_PASSWORD"
+    export VARNAME_TWINE_PASSWORD="EROTEMIC_PYPI_MASTER_TOKEN"
+    export VARNAME_TEST_TWINE_PASSWORD="EROTEMIC_TEST_PYPI_MASTER_TOKEN"
     export VARNAME_PUSH_TOKEN="GITLAB_KITWARE_TOKEN"
+    export VARNAME_TWINE_USERNAME="EROTEMIC_PYPI_MASTER_TOKEN_USERNAME"
+    export VARNAME_TEST_TWINE_USERNAME="EROTEMIC_TEST_PYPI_MASTER_TOKEN_USERNAME"
     export GPG_IDENTIFIER="=Erotemic-CI <erotemic@gmail.com>"
     ' | python -c "import sys; from textwrap import dedent; print(dedent(sys.stdin.read()).strip(chr(10)))" > dev/secrets_configuration.sh
     git add dev/secrets_configuration.sh
@@ -134,10 +136,10 @@ setup_package_environs_gitlab_kitware(){
 setup_package_environs_github_erotemic(){
     echo '
     export VARNAME_CI_SECRET="EROTEMIC_CI_SECRET"
-    export VARNAME_TWINE_USERNAME="TWINE_USERNAME"
-    export VARNAME_TWINE_PASSWORD="TWINE_PASSWORD"
-    export VARNAME_TEST_TWINE_USERNAME="TEST_TWINE_USERNAME"
-    export VARNAME_TEST_TWINE_PASSWORD="TEST_TWINE_PASSWORD"
+    export VARNAME_TWINE_PASSWORD="EROTEMIC_PYPI_MASTER_TOKEN"
+    export VARNAME_TEST_TWINE_PASSWORD="EROTEMIC_TEST_PYPI_MASTER_TOKEN"
+    export VARNAME_TWINE_USERNAME="EROTEMIC_PYPI_MASTER_TOKEN_USERNAME"
+    export VARNAME_TEST_TWINE_USERNAME="EROTEMIC_TEST_PYPI_MASTER_TOKEN_USERNAME"
     export GPG_IDENTIFIER="=Erotemic-CI <erotemic@gmail.com>"
     ' | python -c "import sys; from textwrap import dedent; print(dedent(sys.stdin.read()).strip(chr(10)))" > dev/secrets_configuration.sh
     git add dev/secrets_configuration.sh
@@ -146,9 +148,11 @@ setup_package_environs_github_erotemic(){
 setup_package_environs_github_pyutils(){
     echo '
     export VARNAME_CI_SECRET="PYUTILS_CI_SECRET"
+    export VARNAME_TWINE_PASSWORD="PYUTILS_PYPI_MASTER_TOKEN"
+    export VARNAME_TEST_TWINE_PASSWORD="PYUTILS_TEST_PYPI_MASTER_TOKEN"
+    export VARNAME_TWINE_USERNAME="PYUTILS_PYPI_MASTER_TOKEN_USERNAME"
+    export VARNAME_TEST_TWINE_USERNAME="PYUTILS_TEST_PYPI_MASTER_TOKEN_USERNAME"
     export GPG_IDENTIFIER="=PyUtils-CI <openpyutils@gmail.com>"
-    export VARNAME_TWINE_PASSWORD="PYUTILS_TWINE_PASSWORD"
-    export VARNAME_TWINE_PASSWORD="PYUTILS_TWINE_PASSWORD"
     ' | python -c "import sys; from textwrap import dedent; print(dedent(sys.stdin.read()).strip(chr(10)))" > dev/secrets_configuration.sh
     git add dev/secrets_configuration.sh
 
@@ -161,14 +165,40 @@ setup_package_environs_github_pyutils(){
 upload_github_secrets(){
     load_secrets
     unset GITHUB_TOKEN
-    gh auth login
+    #printf "%s" "$GITHUB_TOKEN" | gh auth login --hostname Github.com --with-token 
+    gh auth login 
     source dev/secrets_configuration.sh
-    gh secret set "$VARNAME_CI_SECRET" -b"${!VARNAME_CI_SECRET}"
-    gh secret set "$VARNAME_TWINE_USERNAME" -b"${!VARNAME_TWINE_USERNAME}"
-    gh secret set "$VARNAME_TWINE_PASSWORD" -b"${!VARNAME_TWINE_PASSWORD}"
-    gh secret set "$VARNAME_TEST_TWINE_PASSWORD" -b"${!VARNAME_TEST_TWINE_PASSWORD}"
-    gh secret set "$VARNAME_TEST_TWINE_USERNAME" -b"${!VARNAME_TEST_TWINE_USERNAME}"
+    gh secret set "TWINE_USERNAME" -b"${!VARNAME_TWINE_USERNAME}"
+    gh secret set "TEST_TWINE_USERNAME" -b"${!VARNAME_TEST_TWINE_USERNAME}"
+    toggle_setx_enter
+    gh secret set "CI_SECRET" -b"${!VARNAME_CI_SECRET}"
+    gh secret set "TWINE_PASSWORD" -b"${!VARNAME_TWINE_PASSWORD}"
+    gh secret set "TEST_TWINE_PASSWORD" -b"${!VARNAME_TEST_TWINE_PASSWORD}"
+    toggle_setx_exit
+}
 
+
+toggle_setx_enter(){
+    # Can we do something like a try/finally?
+    # https://stackoverflow.com/questions/15656492/writing-try-catch-finally-in-shell
+    echo "Enter sensitive area"
+    if [[ -n "${-//[^x]/}" ]]; then
+        __context_1_toggle_setx=1
+    else
+        __context_1_toggle_setx=0
+    fi
+    if [[ "$__context_1_toggle_setx" == "1" ]]; then
+        echo "Setx was on, disable temporarilly"
+        set +x
+    fi
+}
+
+toggle_setx_exit(){
+    echo "Exit sensitive area"
+    # Can we guarentee this will happen?
+    if [[ "$__context_1_toggle_setx" == "1" ]]; then
+        set -x
+    fi
 }
 
 
@@ -208,7 +238,7 @@ upload_gitlab_group_secrets(){
     fi
 
     source dev/secrets_configuration.sh
-    SECRET_VARNAME_ARR=(VARNAME_CI_SECRET VARNAME_TWINE_USERNAME VARNAME_TWINE_PASSWORD VARNAME_TEST_TWINE_PASSWORD VARNAME_TEST_TWINE_USERNAME VARNAME_PUSH_TOKEN)
+    SECRET_VARNAME_ARR=(VARNAME_CI_SECRET VARNAME_TWINE_PASSWORD VARNAME_TEST_TWINE_PASSWORD VARNAME_TWINE_USERNAME VARNAME_TEST_TWINE_USERNAME VARNAME_PUSH_TOKEN)
     for SECRET_VARNAME_PTR in "${SECRET_VARNAME_ARR[@]}"; do
         SECRET_VARNAME=${!SECRET_VARNAME_PTR}
         echo ""
@@ -226,6 +256,8 @@ upload_gitlab_group_secrets(){
         if [[ "$REMOTE_VALUE" == "" ]]; then
             # New variable
             echo "Remove variable does not exist, posting"
+
+            toggle_setx_enter
             curl --request POST --header "PRIVATE-TOKEN: $PRIVATE_GITLAB_TOKEN" "$HOST/api/v4/groups/$GROUP_ID/variables" \
                     --form "key=${SECRET_VARNAME}" \
                     --form "value=${LOCAL_VALUE}" \
@@ -233,11 +265,14 @@ upload_gitlab_group_secrets(){
                     --form "masked=true" \
                     --form "environment_scope=*" \
                     --form "variable_type=env_var" 
+            toggle_setx_exit
         elif [[ "$REMOTE_VALUE" != "$LOCAL_VALUE" ]]; then
             echo "Remove variable does not agree, putting"
             # Update variable value
+            toggle_setx_enter
             curl --request PUT --header "PRIVATE-TOKEN: $PRIVATE_GITLAB_TOKEN" "$HOST/api/v4/groups/$GROUP_ID/variables/$SECRET_VARNAME" \
                     --form "value=${LOCAL_VALUE}" 
+            toggle_setx_exit
         else
             echo "Remote value agrees with local"
         fi
@@ -267,18 +302,25 @@ upload_gitlab_repo_secrets(){
     fi
 
     TMP_DIR=$(mktemp -d -t ci-XXXXXXXXXX)
+    toggle_setx_enter
     curl --header "PRIVATE-TOKEN: $PRIVATE_GITLAB_TOKEN" "$HOST/api/v4/groups" > "$TMP_DIR/all_group_info"
+    toggle_setx_exit
     GROUP_ID=$(cat "$TMP_DIR/all_group_info" | jq ". | map(select(.path==\"$GROUP_NAME\")) | .[0].id")
     echo "GROUP_ID = $GROUP_ID"
 
+    toggle_setx_enter
     curl --header "PRIVATE-TOKEN: $PRIVATE_GITLAB_TOKEN" "$HOST/api/v4/groups/$GROUP_ID" > "$TMP_DIR/group_info"
+    toggle_setx_exit
+    GROUP_ID=$(cat "$TMP_DIR/all_group_info" | jq ". | map(select(.path==\"$GROUP_NAME\")) | .[0].id")
     cat "$TMP_DIR/group_info" | jq
 
     PROJECT_ID=$(cat "$TMP_DIR/group_info" | jq ".projects | map(select(.path==\"$PROJECT_NAME\")) | .[0].id")
     echo "PROJECT_ID = $PROJECT_ID"
 
     # Get group-level secret variables
+    toggle_setx_enter
     curl --header "PRIVATE-TOKEN: $PRIVATE_GITLAB_TOKEN" "$HOST/api/v4/projects/$PROJECT_ID/variables" > "$TMP_DIR/project_vars"
+    toggle_setx_exit
     cat "$TMP_DIR/project_vars" | jq '.[] | .key'
     if [[ "$?" != "0" ]]; then
         echo "Failed to access project level variables. Probably a permission issue"
@@ -286,7 +328,7 @@ upload_gitlab_repo_secrets(){
 
     LIVE_MODE=1
     source dev/secrets_configuration.sh
-    SECRET_VARNAME_ARR=(VARNAME_CI_SECRET VARNAME_TWINE_USERNAME VARNAME_TWINE_PASSWORD VARNAME_TEST_TWINE_PASSWORD VARNAME_TEST_TWINE_USERNAME VARNAME_PUSH_TOKEN)
+    SECRET_VARNAME_ARR=(VARNAME_CI_SECRET VARNAME_TWINE_PASSWORD VARNAME_TEST_TWINE_PASSWORD VARNAME_TWINE_USERNAME VARNAME_TEST_TWINE_USERNAME VARNAME_PUSH_TOKEN)
     for SECRET_VARNAME_PTR in "${SECRET_VARNAME_ARR[@]}"; do
         SECRET_VARNAME=${!SECRET_VARNAME_PTR}
         echo ""
@@ -384,6 +426,11 @@ export_encrypted_code_signing_keys(){
     git add dev/gpg_owner_trust
     git add dev/public_gpg_key
 }
+
+
+# See the xcookie module gitlab python API
+#gitlab_set_protected_branches(){
+#}
 
 
 _test_gnu(){