+++ CHANGELOG.md
@@ -4,7 +4,17 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.6.5 - Unreleased
+## Version 0.7.0 - Unreleased
+
+### Changed
+
+* Flag variables are now accepted on the command line. Variables marked as `isflag` can be specified by either a flag `--varname` or as a key/value pair `--varname=True`.
+
+* The config object now allows aliases to be used in setitem and getitem calls.
+  This should be used sparingly.
+
+
+## Version 0.6.5 - Released 2022-08-25
 
 ### Changed
 * Config.load can now handle aliased input.